Made skipping N2 GUI test for Python 2.7 more transparent
